requirements.txt
torch
transformers
biopython
obonet
scikit-learn
numpy
pandas
tqdm
scikit-multilearn